+++ components/utils.js
@@ -1,7 +1,8 @@
-import { useState, useEffect } from 'react';
+import {useEffect, useState} from 'react';
 import axios from 'axios';
-import { Dimensions } from 'react-native';
+import {Dimensions} from 'react-native';
 import images from "./imageAssets";
+import {createSelector} from "reselect";
 
 export const formatDate = (dateString) => {
     if (!dateString) return 'Date invalide';
@@ -19,8 +20,7 @@ export const fetchFullShowInfo = async (showId) => {
     try {
         const response = await axios.get('https://eurojourney.azurewebsites.net/api/shows');
         const shows = response.data;
-        const specificShow = shows.find(show => show.id === showId);
-        return specificShow;
+        return shows.find(show => show.id === showId);
     } catch (error) {
         console.error('Erreur lors de la récupération des informations complètes du spectacle:', error);
         return null;
@@ -72,3 +72,83 @@ export const importImage = (imageName) => {
         return images['default.jpg'];
     }
 };
+
+export const attractionNames = [
+    'Disneyland Railroad Discoveryland Station',
+    'Disneyland Railroad Fantasyland Station',
+    'Disneyland Railroad Main Street Station',
+    'Disneyland Railroad',
+    'Orbitron®',
+    'Meet Mickey Mouse',
+    'Frontierland Playground',
+    'Disneyland Railroad Frontierland Depot',
+    'Pirate Galleon',
+    'Indiana Jones™ and the Temple of Peril',
+    'La Cabane des Robinson',
+    'Big Thunder Mountain',
+    "Mad Hatter's Tea Cups",
+    'Les Voyages de Pinocchio',
+    'Casey Jr. – le Petit Train du Cirque',
+    'Phantom Manor',
+    'Star Wars Hyperspace Mountain',
+    'Star Tours: The Adventures Continue*',
+    'Thunder Mesa Riverboat Landing',
+    "Alice's Curious Labyrinth",
+    "Buzz Lightyear Laser Blast",
+    'Main Street Vehicles',
+    "Peter Pan's Flight",
+    'Princess Pavilion',
+    'Dumbo the Flying Elephant',
+    "Le Passage Enchanté d'Aladdin",
+    'Autopia, presented by Avis',
+    'Le Carrousel de Lancelot ',
+    'Les Mystères du Nautilus',
+    'La Tanière du Dragon',
+    "Rustler Roundup Shootin' Gallery",
+    'Adventure Isle',
+    'Welcome to Starport: A Star Wars Encounter',
+    "Blanche-Neige et les Sept Nains®",
+    "Mickey’s PhilharMagic",
+    'Pirates of the Caribbean',
+    '"it\'s a small world"',
+    "Le Pays des Contes de Fées",
+    "Pirates' Beach",
+    "Avengers Assemble: Flight Force",
+    "Cars ROAD TRIP",
+    "Spider-Man W.E.B. Adventure",
+    "Cars Quatre Roues Rallye",
+    "Toy Soldiers Parachute Drop",
+    "RC Racer",
+    "The Twilight Zone Tower of Terror",
+    "Crush's Coaster",
+    "Ratatouille: The Adventure",
+    "Slinky® Dog Zigzag Spin",
+    "Les Tapis Volants - Flying Carpets Over Agrabah®"
+];
+export const attractionImages = attractionNames.reduce((acc, name) => {
+    const imageName = formatImageName(name);
+    acc[name] = importImage(imageName);
+    return acc;
+}, {});
+
+export const getRawRideData = state => state.attractions.rawRideData;
+export const getSearchTerm = state => state.attractions.searchTerm;
+export const getFilters = state => state.attractions.filters;
+
+export const getFilteredRideData = createSelector(
+    [getRawRideData, getSearchTerm, getFilters],
+    (rawRideData, searchTerm, filters) => {
+        if (!rawRideData) return [];
+
+        const filteredData = rawRideData.filter(ride => (
+            (!filters.hideClosedRides || ride.status !== 'CLOSED') &&
+            (!filters.showShortWaitTimesOnly || ride.waitTime < 40) &&
+            (filters.selectedPark === 'all' || ride.parkId === filters.selectedPark) &&
+            (filters.selectedType === 'all' || ride.type === filters.selectedType) &&
+            (filters.selectedLand === 'all' || ride.land === filters.selectedLand) &&
+            ride.name.toLowerCase().includes(searchTerm.toLowerCase())
+        ));
+
+        return filteredData.sort((a, b) => a.waitTime - b.waitTime);
+    }
+);
